requirements.txt
streamlit
pandas
matplotlib
geopy
moviepy
pillow
keplergl
streamlit-image-comparison